+++ android-sdk-ui/src/com/appboy/AppboyGcmReceiver.java
@@ -43,6 +43,10 @@ public final class AppboyGcmReceiver extends BroadcastReceiver {
       AppboyNotificationUtils.handleCancelNotificationAction(context, intent);
     } else if (Constants.APPBOY_ACTION_CLICKED_ACTION.equals(action)) {
       AppboyNotificationActionUtils.handleNotificationActionClicked(context, intent);
+    } else if (Constants.APPBOY_STORY_TRAVERSE_CLICKED_ACTION.equals(action)) {
+      handleAppboyGcmReceiveIntent(context, intent);
+    } else if (Constants.APPBOY_STORY_CLICKED_ACTION.equals(action)) {
+      AppboyNotificationUtils.handlePushStoryPageClicked(context, intent);
     } else if (Constants.APPBOY_PUSH_CLICKED_ACTION.equals(action)) {
       AppboyNotificationUtils.handleNotificationOpened(context, intent);
     } else {
@@ -124,6 +128,11 @@ public final class AppboyGcmReceiver extends BroadcastReceiver {
         AppboyConfigurationProvider appConfigurationProvider = new AppboyConfigurationProvider(context);
 
         IAppboyNotificationFactory appboyNotificationFactory = AppboyNotificationUtils.getActiveNotificationFactory();
+
+        if (gcmExtras.containsKey(Constants.APPBOY_PUSH_STORY_KEY) && !gcmExtras.containsKey(Constants.APPBOY_PUSH_STORY_IS_NEWLY_RECEIVED)) {
+          gcmExtras.putBoolean(Constants.APPBOY_PUSH_STORY_IS_NEWLY_RECEIVED, true);
+        }
+
         Notification notification = appboyNotificationFactory.createNotification(appConfigurationProvider, context, gcmExtras, appboyExtras);
 
         if (notification == null) {